産地から選ぶをコンポーネント分割&タグを画面サイズによって折り返すように変更
